+++ HISTORY.md
@@ -1,7 +1,8 @@
 # History
 
 ## Unreleased
-* Consider Vary with an ignored header (via `ignored_parameters`) to always be a cache miss
+* If a request contains a header that is in both `ignored_parameters` and `Vary` (for example,
+`Authorization` + `Vary: Authorization`), always consider it a cache miss
 * Ignore + redact some additional common authentication headers + params by default
 
 ## 1.3.1 (2025-03-03)